issue #162 fix: DateRangePicker - Activity 호출 코드 Event(Callback)로 변경 / Composable 함수들 SearchScreen 파일에 분리
